Fixed a bug with the lastUpdated storage and retrieval in OrganisationUnitCall.
The problem was that serverDate was stored for each OrganisationUnit tree root.
Now it is only stored on successful sync completion.

Also the lastUpdated was retrieved multiple times, when in reality,
it was only necessary to be fetched from the database once for the call.

Protected against the cursor being null (when the table doesn't exist) in the getLastUpdated.